Add files via upload
Across 19 countries and a total sales volume of 9,760,222 million dollars from 2018 to 2020, the USA dominates with the highest sales, reaching 3,355,576 million dollars during this period, solidifying its position as the country with the highest sales.

Sales trends reveal a notable increase from 3.4 million dollars in 2018 to the peak of 4.7 million dollars in 2019. However, a decline is observed in 2020, with total sales dropping to 1.7 million dollars.

Remarkably, the company's top customer is Euro Shopping channel, with transactions totaling 912,294 thousand dollars, reflecting the highest purchase. In contrast, Board and Toys exhibit lower purchasing activity, amounting to just 9,129 thousand dollars within the same period.